modules/cameraCapture/requirements.txt
azure-iot-device~=2.7.0
requests
opencv-python==4.6.0.66